correcciones en abms, dao, etc..

+++ src/modelo/Local.java
@@ -58,7 +58,7 @@ public class Local {
 		return latitud;
 	}
 
-	public void setLatitud(float latitud) {
+	public void setLatitud(double latitud) {
 		this.latitud = latitud;
 	}
 
@@ -66,7 +66,7 @@ public class Local {
 		return longitud;
 	}
 
-	public void setLongitud(float longitud) {
+	public void setLongitud(double longitud) {
 		this.longitud = longitud;
 	}
